reafactor: Box컴포넌트 관련 코드 제거 및 Layout 코드 추가
