+++ apps/client/app/global.css
@@ -11,6 +11,7 @@ body {
   font-family: inherit;
   line-height: inherit;
   margin: 0;
+  background-color: #fafafa;
 }
 h1,
 h2,
