+++ apps/client/app/layout.tsx
@@ -1,17 +1,14 @@
 import './global.css';
+import { PropsWithChildren } from 'react';
 
 export const metadata = {
   title: 'Welcome to client',
   description: 'Generated by create-nx-workspace',
 };
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export default function RootLayout({ children }: PropsWithChildren) {
   return (
-    <html lang="en">
+    <html lang="ko">
       <body>{children}</body>
     </html>
   );
