--- libs/ui/src/components/Box/Box.spec.tsx
@@ -1,9 +1,0 @@
-import { render } from '@testing-library/react';
-import { Box } from './Box';
-
-describe('Box', () => {
-  it('renders', () => {
-    const { baseElement } = render(<Box />);
-    expect(baseElement).toBeTruthy();
-  });
-});
