--- libs/ui/src/components/Box/index.ts
@@ -1,6 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from './Box.stories';
-export * from './Box';
