+++ libs/ui/src/core/index.ts
@@ -2,6 +2,4 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from './layoutUtilityStyle.css';
-export * from './styleProperties';
 export * from './styleToken.css';
