--- libs/ui/src/core/styleProperties.ts
@@ -1,38 +1,0 @@
-import { range } from '@oseek/lib/utils/array';
-
-export const COLOR_PROPERTIES = {
-  primary: '#3673EE',
-  secondary: '#EE3636',
-
-  grayPrimary: '#858B93',
-  graySecondary: '#D6D8DB',
-  grayLight: '#F6F6F6',
-
-  black: '#131313',
-  white: '#FFFFFF',
-};
-
-export const FONT_SIZE_PROPERTIES = {
-  heading1: '22px',
-  subtitle1: '16px',
-  subtitle2: '14px',
-  body: '16px',
-};
-
-export const FONT_WEIGHT_PROPERTIES = {
-  regular: '400',
-  bold: '700',
-};
-
-export const SIZE_PROPERTIES = {
-  statusBarHeight: '24px',
-  headerHeight: '56px',
-  iconSize: '24px',
-  screenMinWidth: '360px',
-  screenMaxWidth: '820px',
-};
-
-export const SPACE_PROPERTIES = range(128).reduce((acc: Record<string, string>, curr) => {
-  acc[curr] = `${curr}px`;
-  return acc;
-}, {})
